src/app/profile/profile.component.ts
import { Component, OnInit } from '@angular/core';
import { AuthService } from '../auth.service'; 
import { Router } from '@angular/router';

@Component({
  selector: 'app-profile',
  templateUrl: './profile.component.html',
  styleUrls: ['./profile.component.css']
})
export class ProfileComponent implements OnInit {

  userEmail: string | null = null;

  constructor(private authService: AuthService, private router: Router) {}

  ngOnInit() {
    this.userEmail = this.authService.getCurrentUserEmail();
  }

  logout() {
    this.authService.logout().then(() => {
      this.router.navigate(['']);
    });
  }

}
